:sparkles: #4 - conflict 처리 및 config 위치 변경
