--- src/main/kotlin/com/teamsparta/tikitaka/infra/security/PasswordEncoderConfig.kt
@@ -1,15 +1,0 @@
-package com.teamsparta.tikitaka.infra.security
-
-import org.springframework.context.annotation.Bean
-import org.springframework.context.annotation.Configuration
-import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder
-import org.springframework.security.crypto.password.PasswordEncoder
-
-@Configuration
-class PasswordEncoderConfig {
-
-    @Bean
-    fun passwordEncoder(): PasswordEncoder {
-        return BCryptPasswordEncoder()
-    }
-}
