--- src/main/kotlin/com/teamsparta/tikitaka/infra/security/SecurityConfig.kt
@@ -1,45 +1,0 @@
-package com.teamsparta.tikitaka.infra.security
-
-import com.teamsparta.tikitaka.infra.security.jwt.JwtAuthenticationFilter
-import org.springframework.context.annotation.Bean
-import org.springframework.context.annotation.Configuration
-import org.springframework.security.config.annotation.method.configuration.EnableMethodSecurity
-import org.springframework.security.config.annotation.web.builders.HttpSecurity
-import org.springframework.security.config.annotation.web.configuration.EnableWebSecurity
-import org.springframework.security.web.AuthenticationEntryPoint
-import org.springframework.security.web.SecurityFilterChain
-import org.springframework.security.web.access.AccessDeniedHandler
-import org.springframework.security.web.authentication.UsernamePasswordAuthenticationFilter
-
-@Configuration
-@EnableWebSecurity
-@EnableMethodSecurity
-class SecurityConfig(
-    private val jwtAuthenticationFilter: JwtAuthenticationFilter,
-    private val authenticationEntryPoint: AuthenticationEntryPoint,
-    private val accessDeniedHandler: AccessDeniedHandler
-) {
-    @Bean
-    fun filterChain(http: HttpSecurity): SecurityFilterChain {
-        return http
-            .httpBasic { it.disable() }
-            .formLogin { it.disable() }
-            .csrf { it.disable() }
-            .authorizeHttpRequests {
-                it.requestMatchers(
-                    "/swagger-ui/**",
-                    "/v3/api-docs/**",
-                    "/api/v1/users/log-in",
-                    "/api/v1/users/sign-up",
-                    "/error"
-                ).permitAll()
-                    .anyRequest().authenticated()
-            }
-            .addFilterBefore(jwtAuthenticationFilter, UsernamePasswordAuthenticationFilter::class.java)
-            .exceptionHandling {
-                it.authenticationEntryPoint(authenticationEntryPoint)
-                it.accessDeniedHandler(accessDeniedHandler)
-            }
-            .build()
-    }
-}
